requirements.txt
# Core dependencies - versions testées et compatibles
streamlit>=1.29.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.6.1

# Vector store and embeddings
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.17.4
python-frontmatter>=1.0.1
markdown>=3.5.1

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.24.3,<2.0.0
pandas>=2.0.3

# Optional
watchdog>=3.0.0
pymdown-extensions>=10.7
